Rename server_settings_log_settings to server_settings_logging
