+++ examples/resources/pingfederate_server_settings_logging/resource.tf
@@ -1,0 +1,32 @@
+resource "pingfederate_server_settings_logging" "logSettings" {
+  log_categories = [
+    {
+      id      = "core"
+      enabled = true
+    },
+    {
+      id      = "policytree"
+      enabled = false
+    },
+    {
+      id      = "trustedcas"
+      enabled = false
+    },
+    {
+      id      = "xmlsig"
+      enabled = false
+    },
+    {
+      id      = "requestheaders"
+      enabled = true
+    },
+    {
+      id      = "requestparams"
+      enabled = false
+    },
+    {
+      id      = "restdatastore"
+      enabled = false
+    },
+  ]
+}
