+++ internal/provider/provider.go
@@ -709,6 +709,7 @@ func (p *pingfederateProvider) DataSources(_ context.Context) []func() datasourc
 		serversettings.ServerSettingsDataSource,
 		serversettingsgeneralsettings.ServerSettingsGeneralDataSource,
 		serversettingsgeneralsettings.ServerSettingsGeneralSettingsDataSource,
+		serversettingslogsettings.ServerSettingsLoggingDataSource,
 		serversettingslogsettings.ServerSettingsLogSettingsDataSource,
 		serversettingssystemkeys.ServerSettingsSystemKeysDataSource,
 		sessionapplicationsessionpolicy.SessionApplicationPolicyDataSource,
@@ -784,6 +785,7 @@ func (p *pingfederateProvider) Resources(_ context.Context) []func() resource.Re
 		serversettings.ServerSettingsResource,
 		serversettingsgeneralsettings.ServerSettingsGeneralResource,
 		serversettingsgeneralsettings.ServerSettingsGeneralSettingsResource,
+		serversettingslogsettings.ServerSettingsLoggingResource,
 		serversettingslogsettings.ServerSettingsLogSettingsResource,
 		serversettingssystemkeys.ServerSettingsSystemKeysResource,
 		sessionapplicationsessionpolicy.SessionApplicationPolicyResource,
